Moved seed extractor storage into its own component as to decouple it from upstream
