+++ Content.Client/RussStation/Botany/Systems/SeedExtractorStorageSystem.cs
@@ -2,4 +2,4 @@ using Content.Shared.RussStation.Botany.Systems;
 
 namespace Content.Client.RussStation.Botany.Systems;
 
-public sealed class SeedExtractorSystem : SharedSeedExtractorSystem;
+public sealed class SeedExtractorStorageSystem : SharedSeedExtractorStorageSystem;
