+++ Content.Server/Botany/Components/SeedExtractorComponent.cs
@@ -1,0 +1,20 @@
+using Content.Shared.Whitelist;
+using Robust.Shared.GameStates;
+
+namespace Content.Server.Botany.Components;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class SeedExtractorComponent : Component
+{
+    /// <summary>
+    /// The minimum amount of seed packets dropped when extracting from produce.
+    /// </summary>
+    [DataField("baseMinSeeds"), ViewVariables(VVAccess.ReadWrite)]
+    public int BaseMinSeeds = 1;
+
+    /// <summary>
+    /// The maximum amount of seed packets dropped when extracting from produce.
+    /// </summary>
+    [DataField("baseMaxSeeds"), ViewVariables(VVAccess.ReadWrite)]
+    public int BaseMaxSeeds = 3;
+}
