+++ Content.Shared/RussStation/Botany/Components/SeedExtractorStorageComponent.cs
@@ -1,0 +1,20 @@
+using Content.Shared.Whitelist;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.RussStation.Botany.Components;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class SeedExtractorStorageComponent : Component
+{
+    /// <summary>
+    /// The ID of the container used to store seed packets placed inside the extractor.
+    /// </summary>
+    [DataField]
+    public string SeedContainerId = "seed_extractor_seeds";
+
+    /// <summary>
+    /// Whitelist controlling which items the extractor will accept.
+    /// </summary>
+    [DataField]
+    public EntityWhitelist? Whitelist;
+}
