+++ Resources/Locale/en-US/@RussStation/botany/components/seed-extractor-storage-component.ftl
@@ -1,0 +1,7 @@
+## Entity
+
+seed-extractor-menu-title = Seed Extractor
+seed-extractor-menu-search = Search...
+seed-extractor-menu-header-name = Name
+seed-extractor-menu-take = Take
+seed-extractor-dump-verb = Deposit into {$unit}
