+++ Resources/Locale/en-US/botany/components/seed-extractor-component.ftl
@@ -2,11 +2,3 @@
 
 seed-extractor-component-interact-message = You extract some seeds from the { THE($name) }.
 seed-extractor-component-no-seeds = { CAPITALIZE(THE($name)) } has no seeds!
-
-## HONK START
-seed-extractor-menu-title = Seed Extractor
-seed-extractor-menu-search = Search...
-seed-extractor-menu-header-name = Name
-seed-extractor-menu-take = Take
-seed-extractor-dump-verb = Deposit into {$unit}
-## HONK END
